Organize tests: co-locate under __tests__, group unit vs integration
Test files move from flat module-sibling placement into per-area
__tests__ directories, grouped by subject:

  src/utils/__tests__/unit/         ai, bet, players, ui, cards.evaluation
  src/utils/__tests__/integration/  showdown, showdown.snapshot,
                                    sidePots.matrix, gameflow.trace
  src/components/__tests__/         components.snapshot
  src/testUtils/__tests__/          factories
  src/App.test.js                   unchanged (beside App.jsx, CRA style)

cards.showdown.test.js and cards.snapshot.test.js are renamed to
showdown.test.js / showdown.snapshot.test.js — their "cards." prefix
only reflected their old location next to cards.js; both exercise the
cross-module showdown pipeline. Snapshot files moved and renamed in
lockstep; all 41 snapshots pass in CI mode without a single rewrite,
proving keys survived the reorganization. Doc references updated.
